+++ src/Game/PositionPredictionSystem.h
@@ -47,7 +47,7 @@ namespace game
             uint32_t timestamp;
             math::Vector position;
             float rotation;
-            short parent_transform;
+            uint32_t parent_transform;
         };
 
         struct PredictionData
